+++ tests/common/auth.ts
