+++ tests/e2e-guest/smoke.spec.ts
@@ -1,6 +1,22 @@
-import { expect, test, trackStep } from "../testBase";
+import { expect, test, trackStep } from "../common/testBase";
 
 test.describe("Smoke @smoke", () => {
+  test("home page renders grid/list of movies", async ({
+    homePage,
+    reportGenerator,
+    testCredentials,
+  }) => {
+    await trackStep(
+      "Verify home page renders grid/list of movies correctly",
+      async () => {
+        await homePage.goTo(testCredentials.baseUrl);
+        await expect(homePage.title).toBeVisible();
+        await expect(homePage.grid).toBeVisible();
+      },
+      reportGenerator.testResult,
+    );
+  });
+  
   test("verify login form displays correctly", async ({
     homePage,
     loginPage,
